+++ choapi/DTOs/LocationDTO.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
 
-        public decimal? Latitude { get; set; } = null;
+        public double? Latitude { get; set; } = null;
 
-        public decimal? Longitude { get; set; } = null;
+        public double? Longitude { get; set; } = null;
     }
 }
