+++ choapi/DTOs/UserDto.cs
@@ -16,9 +16,9 @@
 
         public string? Display_Name { get; set; } = null;
 
-        public decimal? Latitude { get; set; } = null;
+        public double? Latitude { get; set; } = null;
 
-        public decimal? Longitude { get; set; } = null;
+        public double? Longitude { get; set; } = null;
 
         public IFormFile? File { get; set; }
     }
