+++ choapi/Models/Establishment.cs
@@ -17,9 +17,9 @@ namespace choapi.Models
 
         public string? Plan { get; set; } = null;
 
-        public decimal? Latitude { get; set; } = null;
+        public double? Latitude { get; set; } = null;
 
-        public decimal? Longitude { get; set; } = null;
+        public double? Longitude { get; set; } = null;
 
         public bool? Is_Promoted { get; set; } = null;
